v0.0003 pre-alpha
the object oriented update!

recoded clacson to be object oriented and a bit more legibly written, to the best of my ability at present. moved debug toggle to the config file. renamed the ahk and config file. made it so clacson will terminate when the gui is closed. probably other stuff too.